% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_frame.R
\name{column_names.lazy_frame}
\alias{column_names.lazy_frame}
\title{Get column names from a lazy data frame}
\usage{
\method{column_names}{lazy_frame}(lazy_frame)
}
\arguments{
\item{lazy_frame}{lazy frame}
}
\description{
Get column names from a lazy data frame
}
