% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_column.R
\name{lazy_column_test}
\alias{lazy_column_test}
\title{Test a custom implemented lazy column}
\usage{
lazy_column_test(lazy_col)
}
\arguments{
\item{lazy_col}{a custom lazy_column}
}
\value{
TRUE if the teste were succesfull, FALSE otherwise
}
\description{
Test a custom implemented lazy column
}
