% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_table.R
\name{lazy_table}
\alias{lazy_table}
\title{Construct a lazy table}
\usage{
lazy_table(lazy_table_impl)
}
\arguments{
\item{lazy_table_impl}{implementation of a lazy_table class}
}
\value{
a lazy table
}
\description{
Construct a lazy table by providing a `lazy_table_impl` class that can be used to access data. By providing the
interface class, data can be generated on the fly or retrieved from out-of-memory locations such as a disk.
}
