% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_wrap.R
\name{lazy_wrap}
\alias{lazy_wrap}
\title{wrap a table or vector}
\usage{
lazy_wrap(obj)
}
\arguments{
\item{obj}{a data.frame, tibble, data.table or vector object}
}
\value{
a lazy table or lazy vector
}
\description{
wrap a table or vector
}
