% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{op1}
\alias{op1}
\title{Wraps a single argument operator or function}
\usage{
op1(method, vec1, index, length = NULL)
}
\arguments{
\item{method}{method to wrap}

\item{vec1}{a vector, must have length equal to 1 or the length of vec2}

\item{index}{an integer vector specifying the indices to use from the vector, a single integer specifying
the starting index position of the subset or NULL. If a single integer is used, length should be equal to the total number
of elements. If NULL, the full column will be read.}

\item{length}{total number of elements required or NULL if parameter index is set to a integer vector}
}
\value{

}
\description{
Wraps a single argument operator or function
}
