% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{parse_tree}
\alias{parse_tree}
\title{wrap a table or vector}
\usage{
parse_tree(lazy_tbl, e)
}
\arguments{
\item{obj}{a data.frame, tibble, data.table or vector object}
}
\value{
a lazy table or vector
}
\description{
wrap a table or vector
}
