% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_column.R
\name{read_column.lazy_column}
\alias{read_column.lazy_column}
\title{Default lazy column implementation}
\usage{
read_column.lazy_column(lazy_col, index, length)
}
\arguments{
\item{lazy_col}{a custom lazy column}

\item{index}{an integer vector specifying the indices to use from the vector or a single integer specifying
the starting index position of the subset. If a single integer is used, length should be equal to the total number
of elements.}

\item{length}{total number of elements required or NULL if parameter index is set to a integer vector}
}
\description{
Default lazy column implementation
}
