% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_frame.R
\name{read_row_index.lazy_frame}
\alias{read_row_index.lazy_frame}
\title{Read rows from a lazy data frame}
\usage{
\method{read_row_index}{lazy_frame}(lazy_frame, col_name, index)
}
\arguments{
\item{lazy_frame}{lazy frame}

\item{col_name}{column name}

\item{index}{index values}
}
\description{
Read rows from a lazy data frame
}
