% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_table.R
\name{read_row_index}
\alias{read_row_index}
\alias{read_row_range}
\title{Read data from a lazy column}
\usage{
read_row_index(lazy_table_impl, col_name, index)

read_row_range(lazy_table_impl, col_name, from, length)
}
\arguments{
\item{lazy_table_impl}{a custom object with a lazy table API such as a `lazy_frame` (see method `lazy_frame()`)}

\item{col_name}{lazy table column to retrieve data from}

\item{index}{integer vector specifying the index to use from the vector, a single integer specifying
the starting index position of the subset or NULL. If a single integer is used, length should be equal to
the total number of elements. If NULL, the full column will be read.}

\item{from}{starting element}

\item{length}{length of the vector returned}

\item{lazy_frame}{an object generated with `lazy_frame()`}
}
\value{
a subset of a single lazy table column vector

a subset of a single lazy table column vector
}
\description{
Read data from a lazy column

Read data from a lazy column
}
