% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_frame.R
\name{read_row_range.lazy_frame}
\alias{read_row_range.lazy_frame}
\title{Read a row range from a lazy data frame}
\usage{
\method{read_row_range}{lazy_frame}(lazy_frame, col_name, from, length)
}
\arguments{
\item{lazy_frame}{lazy frame}

\item{col_name}{column name}

\item{from}{starting row}

\item{length}{length of selection}
}
\description{
Read a row range from a lazy data frame
}
