% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{select}
\alias{rename}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}}
}}

