% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{rename.lazy_table}
\alias{rename.lazy_table}
\title{dplyr rename}
\usage{
\method{rename}{lazy_table}(.data, ...)
}
\arguments{
\item{.data}{object of class lazy_table}

\item{...}{additional parameters}
}
\description{
dplyr rename
}
