% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.lazy_table}
\alias{select.lazy_table}
\title{dplyr select}
\usage{
\method{select}{lazy_table}(.data, ...)
}
\arguments{
\item{.data}{object of class lazy_table}

\item{...}{additional parameters}
}
\description{
dplyr select
}
