% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_column.R
\name{subset_vec}
\alias{subset_vec}
\title{Subset vector using special semantics}
\usage{
subset_vec(vec, index, length)
}
\arguments{
\item{vec}{vector to subset}

\item{index}{an integer vector specifying the indices to use from the vector, a single integer specifying
the starting index position of the subset or NULL. If a single integer is used, length should be equal to the
total number of elements. If NULL, the full column will be read.}

\item{length}{total number of elements required or NULL if parameter index is set to a integer vector}
}
\value{
a subset of vector `vec`
}
\description{
Subset vector using special semantics
}
