% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{wrap a table or vector}
\usage{
wrap(obj)
}
\arguments{
\item{obj}{a data.frame, tibble, data.table or vector object}
}
\value{
a lazy table or vector
}
\description{
wrap a table or vector
}
