% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_table_generics.R
\name{column_names}
\alias{column_names}
\title{Get column names}
\usage{
column_names(x)
}
\arguments{
\item{x}{a custom object with a lazy table API such as a `lazy_frame` (see method `lazy_frame()`)}
}
\value{
character vector of column names
}
\description{
Get column names
}
