% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_column.R
\name{lazy_column}
\alias{lazy_column}
\title{A lazy column}
\usage{
lazy_column(meta)
}
\arguments{
\item{meta}{custom metadata needed to define the column data}
}
\value{
a lazy column
}
\description{
A lazy column
}
