% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_frame.R
\name{lazy_frame}
\alias{lazy_frame}
\title{Define a lazy data frame}
\usage{
lazy_frame(data_frame)
}
\arguments{
\item{data_frame}{a data frame or equivalent such as a `tibble` or `data.table`}
}
\value{
object of class `lazy_frame` that provides an implementation for the `lazy table` API
}
\description{
A lazy frame is an implementation of the `lazy table` interface for data frames. This class serves
as an example for how to implement a `lazy table`. It wraps an existing data frame and implements the API using that
stored table.
}
