% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_table.R
\name{lazy_table}
\alias{lazy_table}
\title{A lazy table}
\usage{
lazy_table(...)
}
\arguments{
\item{...}{lazy table columns defined as a set of named lazy columns}
}
\value{
a lazy table
}
\description{
A lazy table
}
