% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{parse_tree}
\alias{parse_tree}
\title{restructure an expression for use with sub-setting methods `read_column` and `subset_vec`}
\usage{
parse_tree(e, col_symbols)
}
\arguments{
\item{e}{expression to restructure}

\item{col_symbols}{vector with the symbol names of variables that can be subsetted}
}
\value{
a lazy table or vector
}
\description{
restructure an expression for use with sub-setting methods `read_column` and `subset_vec`
}
