% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_table_generics.R
\name{read_row_index}
\alias{read_row_index}
\title{Read data from a lazy column}
\usage{
read_row_index(lazy_frame_impl, col_name, index)
}
\arguments{
\item{lazy_frame_impl}{a custom object with a lazy table API such as a `lazy_frame` (see method `lazy_frame()`)}

\item{col_name}{lazy table column to retrieve data from}

\item{index}{integer vector specifying the index to use from the vector, a single integer specifying
the starting index position of the subset or NULL. If a single integer is used, length should be equal to
the total number of elements. If NULL, the full column will be read.}
}
\value{
a subset of a single lazy table column vector
}
\description{
Read data from a lazy column
}
