% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_table_generics.R
\name{read_row_range}
\alias{read_row_range}
\title{Read data from a lazy column}
\usage{
read_row_range(lazy_frame_impl, col_name, from, length)
}
\arguments{
\item{lazy_frame_impl}{an object generated with `lazy_frame()`}

\item{col_name}{lazy table column to retrieve data from}

\item{from}{starting element}

\item{length}{length of the vector returned}
}
\value{
a subset of a single lazy table column vector
}
\description{
Read data from a lazy column
}
